requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
appdirs==1.4.4
assertpy==1.1
atomicwrites==1.4.0
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
distlib==0.3.1
filelock==3.0.12
idna==2.10
iniconfig==1.1.1
packaging==20.4
pipenv==2020.8.13
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
pytest-ordering==0.6
requests==2.24.0
selenium==3.141.0
six==1.15.0
toml==0.10.1
urllib3==1.25.11
virtualenv==20.1.0
virtualenv-clone==0.5.4
webdriver-manager==3.2.2
xlrd==1.2.0
